@@ -1,5 +1,4 @@
-﻿using Bannerlord.BLSE.Shared;
-using Bannerlord.BLSE.Shared.Utils;
+﻿using Bannerlord.BLSE.Shared.Utils;
 using Bannerlord.BUTR.Shared.Helpers;
 using Bannerlord.ModuleManager;
 
@@ -82,65 +81,160 @@ internal static class ModuleInitializer
         return null;
     }
 
+    private enum HarmonyDiscoveryResult
+    {
+        Discovered,
+        ModuleMissing,
+        ModuleSubModuleMissing,
+        ModuleSubModuleCorrupted,
+        ModuleVersionWrong,
+        ModuleBinariesMissing,
+        ModuleHarmonyMissing,
+        UnknownIssue,
+    }
+
+    private static HarmonyDiscoveryResult TryResolveHarmonyAssembliesFile(AssemblyName assemblyName, out string? path)
+    {
+        path = null;
+        var assemblyNameFull = $"{assemblyName.Name}.dll";
+
+        var configName = new DirectoryInfo(Directory.GetCurrentDirectory()).Name;
+
+        var harmonyModuleFolder = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "../", "../", ModuleInfoHelper.ModulesFolder, "Bannerlord.Harmony"));
+        if (!Directory.Exists(harmonyModuleFolder))
+            return HarmonyDiscoveryResult.ModuleMissing;
+
+        var harmonySubModule = Path.Combine(harmonyModuleFolder, ModuleInfoHelper.SubModuleFile);
+        if (!File.Exists(harmonySubModule))
+            return HarmonyDiscoveryResult.ModuleSubModuleMissing;
+
+        var doc = new XmlDocument();
+        doc.Load(File.Exists(harmonySubModule) ? harmonySubModule : string.Empty);
+        var harmonyModuleInfo = ModuleInfoExtended.FromXml(doc);
+        if (harmonyModuleInfo is null)
+            return HarmonyDiscoveryResult.ModuleSubModuleCorrupted;
+
+        if (new ApplicationVersionComparer().Compare(harmonyModuleInfo.Version, new ApplicationVersion(ApplicationVersionType.Release, 2, 10, 0, 0)) < 0)
+            return HarmonyDiscoveryResult.ModuleVersionWrong;
+
+        var harmonyBinFolder = Path.Combine(harmonyModuleFolder, "bin", configName);
+        if (!Directory.Exists(harmonyBinFolder))
+            return HarmonyDiscoveryResult.ModuleBinariesMissing;
+
+        var assemblyFile = Path.Combine(harmonyBinFolder, assemblyNameFull);
+        if (!File.Exists(assemblyFile))
+            return HarmonyDiscoveryResult.ModuleHarmonyMissing;
+
+        path = File.Exists(assemblyFile) ? assemblyFile : string.Empty;
+        return HarmonyDiscoveryResult.Discovered;
+    }
+    private static HarmonyDiscoveryResult TryResolveHarmonyAssembliesFileFromSteam(AssemblyName assemblyName, out string? path)
+    {
+        path = null;
+        var assemblyNameFull = $"{assemblyName.Name}.dll";
+
+        var configName = new DirectoryInfo(Directory.GetCurrentDirectory()).Name;
+
+        var harmonySteamModuleFolder = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "../", "../", "../", "../", "workshop", "content", "261550", "2859188632"));
+        if (!Directory.Exists(harmonySteamModuleFolder))
+            return HarmonyDiscoveryResult.ModuleMissing;
+
+        var harmonySteamSubModule = Path.Combine(harmonySteamModuleFolder, ModuleInfoHelper.SubModuleFile);
+        if (!File.Exists(harmonySteamSubModule))
+            return HarmonyDiscoveryResult.ModuleSubModuleMissing;
+
+        var doc = new XmlDocument();
+        doc.Load(File.Exists(harmonySteamSubModule) ? harmonySteamSubModule : string.Empty);
+        var harmonyModuleInfo = ModuleInfoExtended.FromXml(doc);
+        if (harmonyModuleInfo is null)
+            return HarmonyDiscoveryResult.ModuleSubModuleCorrupted;
+
+        if (new ApplicationVersionComparer().Compare(harmonyModuleInfo.Version, new ApplicationVersion(ApplicationVersionType.Release, 2, 10, 0, 0)) < 0)
+            return HarmonyDiscoveryResult.ModuleVersionWrong;
+
+        var harmonyBinSteamFolder = Path.Combine(harmonySteamModuleFolder, "bin", configName);
+        if (!Directory.Exists(harmonyBinSteamFolder))
+            return HarmonyDiscoveryResult.ModuleBinariesMissing;
+
+        var assemblySteamFile = Path.Combine(harmonyBinSteamFolder, assemblyNameFull);
+        if (!File.Exists(assemblySteamFile))
+            return HarmonyDiscoveryResult.ModuleHarmonyMissing;
+
+        path = File.Exists(assemblySteamFile) ? assemblySteamFile : string.Empty;
+        return HarmonyDiscoveryResult.Discovered;
+    }
     private static string? ResolveHarmonyAssembliesFile(AssemblyName assemblyName)
     {
         var assemblyNameFull = $"{assemblyName.Name}.dll";
 
         var configName = new DirectoryInfo(Directory.GetCurrentDirectory()).Name;
-        var harmonyModuleFolder = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "../", "../", ModuleInfoHelper.ModulesFolder, "Bannerlord.Harmony"));
-        // TODO: Proper Steam discovery
-        var harmonySteamModuleFolder = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "../", "../", "../", "../", "workshop", "content", "261550", "2859188632"));
-        if (!Directory.Exists(harmonyModuleFolder) && !Directory.Exists(harmonySteamModuleFolder))
-        {
-            MessageBoxDialog.Show("The Harmony module is missing!\nCan't launch with 'Bannerlord.Harmony' module missing!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
+        var checkSteam = configName == "Win64_Shipping_Client";
 
-        var harmonySubModule = Path.Combine(harmonyModuleFolder, ModuleInfoHelper.SubModuleFile);
-        var harmonySteamSubModule = Path.Combine(harmonySteamModuleFolder, ModuleInfoHelper.SubModuleFile);
-        if (!File.Exists(harmonySubModule) && !File.Exists(harmonySteamSubModule))
-        {
-            MessageBoxDialog.Show($"The Harmony module is corrupted!\nCan't find '{ModuleInfoHelper.SubModuleFile}' in 'Bannerlord.Harmony'!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
+        var genericDiscoveryResult = TryResolveHarmonyAssembliesFile(assemblyName, out var genericHarmonyPath);
+        if (genericDiscoveryResult == HarmonyDiscoveryResult.Discovered)
+            return genericHarmonyPath;
 
-        var doc = new XmlDocument();
-        doc.Load(File.Exists(harmonySubModule) ? harmonySubModule : File.Exists(harmonySteamSubModule) ? harmonySteamSubModule : string.Empty);
-        var harmonyModuleInfo = ModuleInfoExtended.FromXml(doc);
-        if (harmonyModuleInfo is null)
-        {
-            MessageBoxDialog.Show($"The Harmony module is corrupted!\nFailed to read '{ModuleInfoHelper.SubModuleFile}'!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
-        if (new ApplicationVersionComparer().Compare(harmonyModuleInfo.Version, new ApplicationVersion(ApplicationVersionType.Release, 2, 10, 0, 0)) < 0)
-        {
-            MessageBoxDialog.Show("Wrong Harmony module version! At least v2.10.1.x is required!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
+        if (checkSteam && TryResolveHarmonyAssembliesFileFromSteam(assemblyName, out var steamHarmonyPath) == HarmonyDiscoveryResult.Discovered)
+            return steamHarmonyPath;
 
-        var harmonyBinFolder = Path.Combine(harmonyModuleFolder, "bin", configName);
-        var harmonyBinSteamFolder = Path.Combine(harmonySteamModuleFolder, "bin", configName);
-        if (!Directory.Exists(harmonyBinFolder) && !Directory.Exists(harmonyBinSteamFolder))
+        switch (genericDiscoveryResult)
         {
-            MessageBoxDialog.Show($"The Harmony module is corrupted!\nCan't find '{Path.Combine("bin", configName)}' in 'Bannerlord.Harmony'!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
+            case HarmonyDiscoveryResult.ModuleMissing:
+                MessageBoxDialog.Show(@"The Harmony module is missing!
+Can't launch with 'Bannerlord.Harmony' module missing!
 
-        var assemblyFile = Path.Combine(harmonyBinFolder, assemblyNameFull);
-        var assemblySteamFile = Path.Combine(harmonyBinSteamFolder, assemblyNameFull);
-        if (!File.Exists(assemblyFile) && !File.Exists(assemblySteamFile))
-        {
-            MessageBoxDialog.Show($"The Harmony module is corrupted!\nCan't find '{assemblyNameFull}' in 'Bannerlord.Harmony'!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
-            Environment.Exit(1);
-            return null;
-        }
+If the module was installed manually, make sure that the module in installed in 'Modules/Bannerlord.Harmony'!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+            case HarmonyDiscoveryResult.ModuleSubModuleMissing:
+                MessageBoxDialog.Show(@$"The Harmony module is corrupted!
+Can't find '{ModuleInfoHelper.SubModuleFile}' in 'Bannerlord.Harmony'!
 
-        return File.Exists(assemblyFile) ? assemblyFile : File.Exists(assemblySteamFile) ? assemblySteamFile : string.Empty;
+If the module was installed manually, try to do a clean reinstall!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+            case HarmonyDiscoveryResult.ModuleSubModuleCorrupted:
+                MessageBoxDialog.Show(@$"The Harmony module is corrupted!
+Failed to read '{ModuleInfoHelper.SubModuleFile}'!
+
+If the module was installed manually, try to do a clean reinstall!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+            case HarmonyDiscoveryResult.ModuleVersionWrong:
+                MessageBoxDialog.Show(@"The Harmony module is wrong!
+At least v2.10.1.x is required!
+
+If the module was installed manually, find and install the latest version!
+If Vortex is used, try to reinstall manually the latest version!
+If Steam is used, download the latest Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+            case HarmonyDiscoveryResult.ModuleBinariesMissing:
+                MessageBoxDialog.Show(@$"The Harmony module is corrupted!
+Can't find '{Path.Combine("bin", configName)}' in 'Bannerlord.Harmony'!
+
+If the module was installed manually, try to do a clean reinstall!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+            case HarmonyDiscoveryResult.ModuleHarmonyMissing:
+                MessageBoxDialog.Show(@$"The Harmony module is corrupted!
+Can't find '{assemblyNameFull}' in 'Bannerlord.Harmony'!
+
+If the module was installed manually, try to do a clean reinstall!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
+                Environment.Exit(1);
+                return null;
+        }
+        return null;
     }
 
     private static Assembly? ResolveHarmonyAssembly(AssemblyName assemblyName)
@@ -158,7 +252,12 @@ internal static class ModuleInitializer
         var harmonyX = AssemblyDefinition.ReadAssembly(assemblyFile);
         if (harmonyX.Name.Version < new Version(2, 10, 1, 0))
         {
-            MessageBoxDialog.Show("The Harmony module is corrupted!\nWrong 0Harmony.dll version! At least v2.10.1.x is required!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error, 0, 0);
+            MessageBoxDialog.Show(@"The Harmony module is corrupted!
+Wrong 0Harmony.dll version! At least v2.10.1.x is required!
+
+If the module was installed manually, try to do a clean reinstall!
+If Vortex is used, try to reinstall manually!
+If Steam is used, download the Harmony mod from NexusMods!", "Error from BLSE!", MessageBoxButtons.Ok, MessageBoxIcon.Error);
             Environment.Exit(1);
             return null;
         }
